feat: rename utils func
