+++ my-app/src/utils/getLiOptions.ts
